refactor: decoupled client schemas into new business layer (tier 3) files
