+++ frontend/src/components/clients/tabs/LogisticsTab.tsx
@@ -9,7 +9,8 @@ import { db } from "@/firebase/firebase";
 
 import { z } from "zod";
 import { AccordionSection } from "@/components/ui/AccordionSection";
-import { logisticsSchema } from "@/schemas/clientSchema";
+
+import { logisticsSchema } from "@/schema/supplementarySchema";
 import type { ClientDoc } from "@/components/clients/list/ClientList";
 
 // ─── Props ────────────────────────────────────────────────────────────────────
