+++ frontend/src/components/clients/tabs/QuestionnaireTab.tsx
@@ -9,7 +9,7 @@ import { db } from "@/firebase/firebase";
 
 import { z } from "zod";
 import { AccordionSection } from "@/components/ui/AccordionSection";
-import { questionnaireSchema } from "@/schemas/clientSchema";
+import { questionnaireSchema } from "@/schema/supplementarySchema";
 import type { ClientDoc } from "@/components/clients/list/ClientList";
 
 // ─── Props ────────────────────────────────────────────────────────────────────
